requirements.txt
Ctransformers
gradio
langchain
pyyaml
termcolor
requests
jsonlines
openai
chromadb